+++ qb_device_driver/src/qb_device_communication_handler.cpp
@@ -1,7 +1,7 @@
 /***
  *  Software License Agreement: BSD 3-Clause License
  *
- *  Copyright (c) 2016-2018, qbrobotics®
+ *  Copyright (c) 2016-2021, qbrobotics®
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
@@ -31,12 +31,7 @@ using namespace qb_device_communication_handler;
 
 qbDeviceCommunicationHandler::qbDeviceCommunicationHandler()
     : qbDeviceCommunicationHandler(std::make_shared<qb_device_driver::qbDeviceAPI>()) {
-  while (!getSerialPortsAndDevices(1)) {
-    ROS_WARN_STREAM_NAMED("communication_handler", "[CommunicationHandler] is waiting for devices...");
-    ros::Duration(1.0).sleep();
-  }
-
-  spinner_.start();
+  
 }
 
 qbDeviceCommunicationHandler::qbDeviceCommunicationHandler(qb_device_driver::qbDeviceAPIPtr device_api)
@@ -49,11 +44,21 @@ qbDeviceCommunicationHandler::qbDeviceCommunicationHandler(qb_device_driver::qbD
       initialize_device_(node_handle_.advertiseService("/communication_handler/initialize_device", &qbDeviceCommunicationHandler::initializeCallback, this)),
       set_commands_(node_handle_.advertiseService("/communication_handler/set_commands", &qbDeviceCommunicationHandler::setCommandsCallback, this)),
       set_pid_(node_handle_.advertiseService("/communication_handler/set_pid", &qbDeviceCommunicationHandler::setPIDCallback, this)),
+      set_control_mode_(node_handle_.advertiseService("/communication_handler/set_control_mode", &qbDeviceCommunicationHandler::setControlModeCallback, this)),
+      connection_state_publisher_(node_handle_.advertise<qb_device_msgs::ConnectionState>("/communication_handler/connection_state",1)),
+      
       device_api_(device_api) {
-
+  while (!getSerialPortsAndDevices(1)) {
+    ROS_WARN_STREAM_NAMED("communication_handler", "[CommunicationHandler] is waiting for devices...");
+    ros::Duration(1.0).sleep();
+  }
+  check_connection_status_timer_ = node_handle_.createWallTimer(ros::WallDuration(1), &qbDeviceCommunicationHandler::checkConnectionAndPublish, this);
+  spinner_.start();
+  
 }
 
 qbDeviceCommunicationHandler::~qbDeviceCommunicationHandler() {
+  ROS_INFO_STREAM_NAMED("communication_handler", "[CommunicationHandler] Attempt to terminate the process properly...");
   for (auto const &file_descriptor : file_descriptors_) {
     close(file_descriptor.first);
   }
@@ -69,7 +74,7 @@ int qbDeviceCommunicationHandler::activate(const int &id, const bool &command, c
     device_api_->activate(&file_descriptors_.at(connected_devices_.at(id)), id, command);
     failures = std::max(failures, isActive(id, max_repeats, status));
     if (status != command) {
-      ROS_ERROR_STREAM_NAMED("communication_handler", "[CommunicationHandler] device [" << id << "] fails on " << command_prefix << "activation.");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"communication_handler", "[CommunicationHandler] device [" << id << "] fails on " << command_prefix << "activation.");
       return -1;
     }
     ROS_INFO_STREAM_NAMED("communication_handler", "[CommunicationHandler] device [" << id << "] motors have been " << command_prefix << "activated!");
@@ -107,9 +112,10 @@ int qbDeviceCommunicationHandler::close(const std::string &serial_port) {
   for (auto const &device : connected_devices_) {
     if (device.second == serial_port) {
       deactivate(device.first, 3);
-      connected_devices_.erase(device.first);
     }
   }
+  connected_devices_.clear();
+  ROS_INFO_STREAM_NAMED("communication_handler", "# of connected devices " << connected_devices_.size());
   device_api_->close(&file_descriptors_.at(serial_port));
   file_descriptors_.erase(serial_port);
   ROS_INFO_STREAM_NAMED("communication_handler", "[CommunicationHandler] does not handle [" << serial_port << "] anymore.");
@@ -177,6 +183,20 @@ bool qbDeviceCommunicationHandler::getInfoCallback(qb_device_srvs::TriggerReques
   return true;
 }
 
+int qbDeviceCommunicationHandler::getCommands(const int &id, const int &max_repeats, std::vector<short int> &commands) {
+  // the API methods are called at most (i.e. very unlikely) 'max_repeats' times to guarantee the correct identification of a real fault in the communication
+  int failures = 0;
+  commands.resize(2);  // required by 'getCurrents()'
+  while (failures <= max_repeats) {
+    if (device_api_->getCommands(&file_descriptors_.at(connected_devices_.at(id)), id, commands) < 0) {
+      failures++;
+      continue;
+    }
+    break;
+  }
+  return failures;
+}
+
 int qbDeviceCommunicationHandler::getMeasurements(const int &id, const int &max_repeats, std::vector<short int> &currents, std::vector<short int> &positions) {
   // the API methods are called at most (i.e. very unlikely) 'max_repeats' times to guarantee the correct identification of a real fault in the communication
   int failures = 0;
@@ -206,18 +226,22 @@ bool qbDeviceCommunicationHandler::getMeasurementsCallback(qb_device_srvs::GetMe
   response.failures = 0;  // need to return true even if both 'get_currents' and 'get_positions' are set to false
   if (request.get_currents && request.get_positions) {
     if (!request.get_distinct_packages) {
-      response.failures = getMeasurements(request.id, request.max_repeats, response.currents, response.positions);  // blocks while reading
+      response.failures += getMeasurements(request.id, request.max_repeats, response.currents, response.positions);  // blocks while reading
     }
     else {
-      response.failures = getCurrents(request.id, request.max_repeats, response.currents)
+      response.failures += getCurrents(request.id, request.max_repeats, response.currents)
                           + getPositions(request.id, request.max_repeats, response.positions);  // both block while reading
     }
   }
   else if (request.get_currents) {
-    response.failures = getCurrents(request.id, request.max_repeats, response.currents);  // blocks while reading
+    response.failures += getCurrents(request.id, request.max_repeats, response.currents);  // blocks while reading
   }
   else if (request.get_positions) {
-    response.failures = getPositions(request.id, request.max_repeats, response.positions);  // blocks while reading
+    response.failures += getPositions(request.id, request.max_repeats, response.positions);  // blocks while reading
+  }
+
+  if (request.get_commands) {
+    response.failures += getCommands(request.id, request.max_repeats, response.commands);  // blocks while reading
   }
 
   response.stamp = ros::Time::now();
@@ -229,7 +253,7 @@ int qbDeviceCommunicationHandler::getParameters(const int &id, std::vector<int> 
   std::vector<int> input_mode = {-1};
   std::vector<int> control_mode = {-1};
   device_api_->getParameters(&file_descriptors_.at(connected_devices_.at(id)), id, input_mode, control_mode, resolutions, limits);
-  if (!input_mode.front() && !control_mode.front()) {  // both input and control modes equals 0 are required, i.e. respectively USB connected and position controlled
+  if (!input_mode.front() && (!control_mode.front() || control_mode.front() == 4)) {  // both input and control modes equals 0 are required, i.e. respectively USB connected and position controlled (also deflection control is possible with qbmoves)
     return 0;
   }
   return -1;
@@ -269,14 +293,21 @@ int qbDeviceCommunicationHandler::getSerialPortsAndDevices(const int &max_repeat
     // 'serial_protectors_' is not cleared because of the previously acquired lock, do not do it!
     serial_protectors_.insert(std::make_pair(serial_ports.at(i), std::make_unique<std::mutex>()));  // never override
 
+    int at_least_one_device_found = 0;
     std::array<char, 255> devices;
     int devices_number = device_api_->getDeviceIds(&file_descriptors_.at(serial_ports.at(i)), devices);
     for (int j=0; j<devices_number; j++) {
-      if (devices.at(j) == 120) {
-        continue;  // ID 120 is reserved for dummy board which should not be considered as a connected device
+      if (devices.at(j) == 120 || devices.at(j) == 0) {
+        continue;  // ID 120 is reserved for dummy board which should not be considered as a connected device (ID 0 is for sure an error)
       }
       // actually a std::map does not let same-id devices on distinct serial ports
       connected_devices.insert(std::make_pair(static_cast<int>(devices.at(j)), serial_ports.at(i)));
+      at_least_one_device_found++;
+    }
+
+    if (!at_least_one_device_found) {
+      ROS_INFO_STREAM_NAMED("communication_handler", "[CommunicationHandler] no qbrobotics devices found on [" << serial_ports.at(i) << "]");
+      close(serial_ports.at(i));
     }
   }
 
@@ -326,7 +357,7 @@ bool qbDeviceCommunicationHandler::isInOpenMap(const std::string &serial_port) {
 
 int qbDeviceCommunicationHandler::open(const std::string &serial_port) {
   if (!std::regex_match(serial_port, std::regex("/dev/ttyUSB[[:digit:]]+"))) {
-    ROS_ERROR_STREAM_NAMED("communication_handler", "[CommunicationHandler] fails while opening [" << serial_port << "] because it does not match the expected pattern [/dev/ttyUSB*].");
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"communication_handler", "[CommunicationHandler] fails while opening [" << serial_port << "] because it does not match the expected pattern [/dev/ttyUSB*].");
     return -1;
   }
 
@@ -337,7 +368,7 @@ int qbDeviceCommunicationHandler::open(const std::string &serial_port) {
 
   device_api_->open(&file_descriptors_[serial_port], serial_port);  // also create a pair in the map
   if(file_descriptors_.at(serial_port).file_handle == INVALID_HANDLE_VALUE) {
-    ROS_ERROR_STREAM_NAMED("communication_handler", "[CommunicationHandler] fails while opening [" << serial_port << "] and sets errno [" << strerror(errno) << "].");
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"communication_handler", "[CommunicationHandler] fails while opening [" << serial_port << "] and sets errno [" << strerror(errno) << "].");
     // remove file descriptor entry
     file_descriptors_.erase(serial_port);
     return -1;
@@ -354,9 +385,12 @@ bool qbDeviceCommunicationHandler::initializeCallback(qb_device_srvs::Initialize
     serial_locks.push_back(std::unique_lock<std::mutex>(*mutex.second));
   }
 
-  if (request.rescan) {
+  if (request.rescan || !isInConnectedSet(request.id)) {
     // update connected devices
+    ROS_INFO_STREAM_NAMED("communication_handler", "[CommunicationHandler] Rescanning...");
+    check_connection_status_timer_.stop();
     getSerialPortsAndDevices(request.max_repeats);
+    check_connection_status_timer_.start();
   }
   if (!isInConnectedSet(request.id) || !isReliable(isConnected(request.id, request.max_repeats), request.max_repeats)) {
     ROS_ERROR_STREAM_NAMED("communication_handler", "[CommunicationHandler] fails while initializing device [" << request.id << "] because it is not connected.");
@@ -386,6 +420,29 @@ bool qbDeviceCommunicationHandler::initializeCallback(qb_device_srvs::Initialize
   response.message = "Device [" + std::to_string(request.id) + "] initialization succeeds.";
   response.success = true;
   return true;
+}
+
+void qbDeviceCommunicationHandler::checkConnectionAndPublish(const ros::WallTimerEvent &timer_event) {
+  qb_device_msgs::DeviceConnectionInfo device_info_;
+  bool status = false;
+  int failures = 0;
+  const int max_repeats = 3;
+  connection_state_msg_.devices.clear();
+
+  for(auto devices : connected_devices_) {
+    std::lock_guard<std::mutex> serial_lock(*serial_protectors_.at(connected_devices_.at(devices.first)));
+    failures = isActive(devices.first, max_repeats, status);
+    if (failures >= max_repeats) { // too much consecutive failues
+      device_info_.is_active = false;
+    } else {
+      device_info_.is_active = status;
+    }
+    device_info_.id = devices.first;
+    device_info_.port = devices.second;
+    connection_state_msg_.devices.push_back(device_info_);
+  }
+
+  connection_state_publisher_.publish(connection_state_msg_);
 }
 
 int qbDeviceCommunicationHandler::setCommandsAndWait(const int &id, const int &max_repeats, std::vector<short int> &commands) {
@@ -459,4 +516,38 @@ bool qbDeviceCommunicationHandler::setPIDCallback(qb_device_srvs::SetPIDRequest 
   response.failures = setPID(request.id, request.max_repeats, pid);  // blocking function
   response.success = isReliable(response.failures, request.max_repeats);
   return true;
+}
+
+int qbDeviceCommunicationHandler::setControlMode(const int &id, const int &max_repeats, uint8_t &control_id) {
+  int failures = 0;
+  while (failures <= max_repeats) {
+    if (device_api_->setControlMode(&file_descriptors_.at(connected_devices_.at(id)), id, control_id) < 0) {
+      failures++;
+      continue;
+    }
+    break;
+  }
+  return failures;
+}
+
+
+bool qbDeviceCommunicationHandler::setControlModeCallback(qb_device_srvs::SetControlModeRequest &request, qb_device_srvs::SetControlModeResponse &response) {
+  ROS_ERROR_STREAM_COND_NAMED(request.max_repeats < 0, "communication_handler", "Device [" << request.id << "] has request service with non-valid 'max_request' [" << request.max_repeats << "].");
+  if (!isInConnectedSet(request.id)) {
+    response.success = false;
+    return true;
+  }
+  uint8_t control_id;
+  if (request.control == "position") {
+    control_id = 0;
+  } else if (request.control == "deflection") {
+    control_id = 4;
+  } else {
+    ROS_ERROR_STREAM_NAMED("communication_handler","Not valid control mode request.");
+    response.success = false;
+    return true;
+  }
+  response.failures = setControlMode(request.id, request.max_repeats, control_id);  // blocking function
+  response.success = isReliable(response.failures, request.max_repeats);
+  return true; 
 }