+++ qb_device_hardware_interface/src/qb_device_hardware_interface.cpp
@@ -1,7 +1,7 @@
 /***
  *  Software License Agreement: BSD 3-Clause License
  *
- *  Copyright (c) 2016-2018, qbrobotics®
+ *  Copyright (c) 2016-2021, qbrobotics®
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
@@ -50,13 +50,13 @@ int qbDeviceHW::activateMotors() {
     srv.request.max_repeats = device_.max_repeats;
     services_.at("activate_motors").call(srv);
     if (!srv.response.success) {
-      ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot activate device [" << device_.id << "].");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot activate device [" << device_.id << "].");
       return -1;
     }
-    ROS_INFO_STREAM_NAMED("device_hw", "[DeviceHW] device [" << device_.id << "] motors are active!");
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] device [" << device_.id << "] motors are active!");
     return 0;
   }
-  ROS_WARN_STREAM_NAMED("device_hw", "[DeviceHW] service [activate_motors] seems no longer advertised. Trying to reconnect...");
+  ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] service [activate_motors] seems no longer advertised. Trying to reconnect...");
   resetServicesAndWait();
   return -1;
 }
@@ -72,6 +72,23 @@ std::vector<std::string> qbDeviceHW::addNamespacePrefix(const std::vector<std::s
   return namespaced_vector;
 }
 
+void qbDeviceHW::controllerStatusCallback(const control_msgs::JointTrajectoryControllerState &status_msg) {
+  if (status_msg.header.seq < 30 || controller_state_counter_++ < 30) {  // filter out first messages
+    return;
+  }
+
+  if (controller_first_command_publisher_) {  // this will be called just once because the publisher will be killed after the first trigger
+    ROS_INFO_STREAM_NAMED("device_hw", "[DeviceHW] sending the first command references to the device [" << device_.id << "] trajectory controller...");
+    controller_first_command_publisher_.publish(controller_first_point_trajectory_);  // it can be published now because the controller is properly spawned at this point
+
+    std_msgs::Header msg;
+    msg.frame_id = device_.name;
+    controller_startup_sync_publisher_.publish(msg);  // used to sync the control loop setup callback
+  }
+  controller_first_command_publisher_.shutdown();  // terminate the first command publisher
+  controller_state_subscriber_.shutdown();  // suicide because it will be no longer useful
+}
+
 int qbDeviceHW::deactivateMotors() {
   if (services_.at("deactivate_motors")) {
     qb_device_srvs::Trigger srv;
@@ -79,7 +96,7 @@ int qbDeviceHW::deactivateMotors() {
     srv.request.max_repeats = device_.max_repeats;
     services_.at("deactivate_motors").call(srv);
     if (!srv.response.success) {
-      ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot deactivate device [" << device_.id << "].");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot deactivate device [" << device_.id << "].");
       return -1;
     }
     ROS_INFO_STREAM_NAMED("device_hw", "[DeviceHW] device [" << device_.id << "] motors are inactive.");
@@ -90,6 +107,27 @@ int qbDeviceHW::deactivateMotors() {
   return -1;
 }
 
+int qbDeviceHW::getCommands(std::vector<double> &commands) {
+  commands.resize(2);
+
+  // read actual references state from the hardware
+  bool user_get_commands = device_.get_commands;
+  device_.get_commands = true;  // used to retrieve also the commands
+  actuators_.setReliability(device_.max_repeats, getMeasurements(actuators_.positions, actuators_.efforts, actuators_.commands, actuators_.stamp));
+  device_.get_commands = user_get_commands;
+  if (!actuators_.is_reliable) {
+    return -1;
+  }
+
+  // propagate current actuator state to joints
+  transmission_.actuator_to_joint_state.propagate();
+
+  commands.at(0) = actuators_.commands.at(0) * joints_.positions.at(0)/actuators_.positions.at(0);  // commands are not converted inside the transmission
+  commands.at(1) = actuators_.commands.at(1) * joints_.positions.at(1)/actuators_.positions.at(1);  // commands are not converted inside the transmission
+
+  return 0;
+}
+
 std::string qbDeviceHW::getInfo() {
   if (services_.at("get_info")) {
     qb_device_srvs::Trigger srv;
@@ -97,7 +135,7 @@ std::string qbDeviceHW::getInfo() {
     srv.request.max_repeats = device_.max_repeats;
     services_.at("get_info").call(srv);
     if (!srv.response.success) {
-      ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot get info from device [" << device_.id << "].");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot get info from device [" << device_.id << "].");
       return "";
     }
     return srv.response.message;
@@ -108,6 +146,11 @@ std::string qbDeviceHW::getInfo() {
 }
 
 int qbDeviceHW::getMeasurements(std::vector<double> &positions, std::vector<double> &currents, ros::Time &stamp) {
+  std::vector<double> commands;  // not used
+  return getMeasurements(positions, currents, commands, stamp);
+}
+
+int qbDeviceHW::getMeasurements(std::vector<double> &positions, std::vector<double> &currents, std::vector<double> &commands, ros::Time &stamp) {
   if (services_.at("get_measurements")) {
     qb_device_srvs::GetMeasurements srv;
     srv.request.id = device_.id;
@@ -115,9 +158,10 @@ int qbDeviceHW::getMeasurements(std::vector<double> &positions, std::vector<doub
     srv.request.get_currents = device_.get_currents;
     srv.request.get_positions = device_.get_positions;
     srv.request.get_distinct_packages = device_.get_distinct_packages;
+    srv.request.get_commands = device_.get_commands;  // usually false, it is used for async requests only
     services_.at("get_measurements").call(srv);
     if (!srv.response.success) {
-      ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot get measurements from device [" << device_.id << "].");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot get measurements from device [" << device_.id << "].");
       return srv.response.failures;  // positions and currents are not updated
     }
 
@@ -127,6 +171,9 @@ int qbDeviceHW::getMeasurements(std::vector<double> &positions, std::vector<doub
     }
     for (int i=0; i<currents.size() && i<srv.response.currents.size(); i++) {
       currents.at(i) = srv.response.currents.at(i);
+    }
+    for (int i=0; i<commands.size() && i<srv.response.commands.size(); i++) {
+      commands.at(i) = srv.response.commands.at(i);
     }
     stamp = srv.response.stamp;
     return srv.response.failures;
@@ -139,15 +186,15 @@ int qbDeviceHW::getMeasurements(std::vector<double> &positions, std::vector<doub
 bool qbDeviceHW::init(ros::NodeHandle &root_nh, ros::NodeHandle &robot_hw_nh) {
   node_handle_ = robot_hw_nh;
   if (!robot_hw_nh.getParam("device_name", device_.name)) {
-    ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot retrieve 'device_name' from the Parameter Service [" << robot_hw_nh.getNamespace() << "].");
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot retrieve 'device_name' from the Parameter Server [" << robot_hw_nh.getNamespace() << "].");
     return false;
   }
   if (!robot_hw_nh.getParam("device_id", device_.id)) {
-    ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot retrieve 'device_id' from the Parameter Service [" << robot_hw_nh.getNamespace() << "].");
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot retrieve 'device_id' from the Parameter Server [" << robot_hw_nh.getNamespace() << "].");
     return false;
   }
   if (!urdf_model_.initParamWithNodeHandle("robot_description", root_nh)) {
-    ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot retrieve 'robot_description' from the Parameter Service [" << robot_hw_nh.getNamespace() << "].");
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot retrieve 'robot_description' from the Parameter Server [" << robot_hw_nh.getNamespace() << "].");
     return false;
   }
 
@@ -157,13 +204,51 @@ bool qbDeviceHW::init(ros::NodeHandle &root_nh, ros::NodeHandle &robot_hw_nh) {
   joints_.setJoints(robot_hw_nh.param<std::vector<std::string>>("joints", addNamespacePrefix(joints_.names)));
 
   interfaces_.initialize(this, joints_);
-  joint_limits_.initialize(robot_hw_nh, joints_, urdf_model_, interfaces_.joint_position);
+  // initialized later on joint_limits_.initialize(robot_hw_nh, joints_, urdf_model_, interfaces_.joint_position);
   transmission_.initialize(robot_hw_nh.param<std::string>("transmission", "transmission"), actuators_, joints_);
 
-  use_simulator_mode_ = robot_hw_nh.param<bool>("use_simulator_mode", false);
+  use_fake_measurement_mode_ = robot_hw_nh.param<bool>("use_fake_measurement_mode", false);
 
   waitForInitialization();
   ROS_INFO_STREAM(getInfo());
+
+  std::string default_controller;
+  if (!robot_hw_nh.getParam("default_controller", default_controller)) {
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot retrieve 'default_controller' from the Parameter Server [" << robot_hw_nh.getNamespace() << "].");
+    return false;
+  }
+  std::vector<std::string> default_controller_joints;
+  if (!robot_hw_nh.getParam(default_controller + "/joints", default_controller_joints)) {
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot retrieve 'joints' from the controller in the Parameter Server [" << robot_hw_nh.getNamespace() << "].");
+    return false;
+  }
+  std::vector<double> commands;
+  if (getCommands(commands) == -1) {
+    ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot retrieve command references from device [" << device_.id << "].");
+    return false;
+  }
+  // actual joint commands (coming from actual device references) are required while initializing the joint limit interface (to set the starting old command for enforceLimits())
+  joints_.commands.at(0) = commands.at(0);
+  joints_.commands.at(1) = commands.at(1);
+  joints_.commands.at(2) = (commands.at(0) + commands.at(1))/2;
+  joints_.commands.at(3) = std::abs(commands.at(0) - commands.at(1))/2;
+  joint_limits_.initialize(robot_hw_nh, joints_, urdf_model_, interfaces_.joint_position);
+
+  controller_first_command_publisher_ = node_handle_.advertise<trajectory_msgs::JointTrajectory>(default_controller + "/command", 1);
+  controller_state_subscriber_ = node_handle_.subscribe(default_controller + "/state", 1, &qbDeviceHW::controllerStatusCallback, this);
+  trajectory_msgs::JointTrajectoryPoint point;
+  point.positions.resize(commands.size());
+  point.velocities.resize(commands.size());
+  point.accelerations.resize(commands.size());
+  point.effort.resize(commands.size());
+  point.positions.at(0) = joints_.commands.at(2);
+  point.positions.at(1) = joints_.commands.at(3);
+  point.time_from_start = ros::Duration(0.1);
+  controller_first_point_trajectory_.points.push_back(point);
+  controller_first_point_trajectory_.joint_names = default_controller_joints;
+  // cannot publish it yet (the controller has not been spawned yet) first_command_publisher_.publish(first_point_trajectory_);
+
+  controller_startup_sync_publisher_ = root_nh.advertise<std_msgs::Header>("control/startup_sync", 1);  // WARN: root_nh is different!
 
   return true;
 }
@@ -172,13 +257,13 @@ int qbDeviceHW::initializeDevice() {
   if (services_.at("initialize_device")) {
     qb_device_srvs::InitializeDevice srv;
     srv.request.id = device_.id;
-    srv.request.activate = node_handle_.param<bool>("activate_on_initialization", false) && !use_simulator_mode_;
+    srv.request.activate = node_handle_.param<bool>("activate_on_initialization", false) && !use_fake_measurement_mode_;
     srv.request.rescan = node_handle_.param<bool>("rescan_on_initialization", false);
     int max_repeats = node_handle_.param<int>("max_repeats", 3);
     srv.request.max_repeats = max_repeats;
     services_.at("initialize_device").call(srv);
     if (!srv.response.success) {
-      ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot initialize device [" << device_.id << "].");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot initialize device [" << device_.id << "].");
       return -1;
     }
     device_.max_repeats = max_repeats;
@@ -187,6 +272,7 @@ int qbDeviceHW::initializeDevice() {
     device_.get_distinct_packages = node_handle_.param<bool>("get_distinct_packages", false);
     device_.set_commands = node_handle_.param<bool>("set_commands", true);
     device_.set_commands_async = node_handle_.param<bool>("set_commands_async", false);
+    node_handle_.getParam("use_joint_limits", device_.use_joint_limits);
     device_.serial_port = srv.response.info.serial_port;
     device_.position_limits = srv.response.info.position_limits;
     device_.encoder_resolutions = srv.response.info.encoder_resolutions;
@@ -264,7 +350,7 @@ void qbDeviceHW::read(const ros::Time &time, const ros::Duration &period) {
     }
   }
 
-  if (use_simulator_mode_) {
+  if (use_fake_measurement_mode_) {
     actuators_.positions = actuators_.commands;
   }
 
@@ -300,7 +386,7 @@ int qbDeviceHW::setCommands(const std::vector<double> &commands) {
 
     services_.at("set_commands").call(srv);
     if (!srv.response.success) {
-      ROS_ERROR_STREAM_NAMED("device_hw", "[DeviceHW] cannot send commands to device [" << device_.id << "].");
+      ROS_ERROR_STREAM_THROTTLE_NAMED(60 ,"device_hw", "[DeviceHW] cannot send commands to device [" << device_.id << "].");
       return -1;
     }
     return 0;