+++ qb_device/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package qb_device
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.2.0 (2021-08-27)
+------------------
+* Refactor Gazebo plugin to prepare for SoftHand simulation
+
 2.1.1 (2019-10-07)
 ------------------